requirements.txt
# --- Core ---
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# --- Configuration & Validation ---
pydantic>=2.8.0
pydantic-settings>=2.4.0
email-validator>=2.2.0

# --- Database ---
sqlalchemy>=2.0.0
alembic>=1.13.0

# --- Security ---
python-jose>=3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3


# --- HTTP & Utils ---
python-multipart>=0.0.9
httpx>=0.27.0

# LLM
openai>=1.40.0

# Text Extraction
trafilatura>=1.9.0
requests>=2.32.0
youtube-transcript-api>=0.6.2

# Dockerfile runs FastAPI with Gunicorn + Uvicorn workers (production standard)
gunicorn==22.0.0


# using SQLAlchemy with Postgres
psycopg2-binary==2.9.9
